+++ src/logs/coralogix.rs
@@ -126,9 +126,15 @@ struct JsonMessage {
     kinesis_event_id: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "kinesis.event.name")]
     kinesis_event_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", rename = "kinesis.event.source")]
+    #[serde(
+        skip_serializing_if = "Option::is_none",
+        rename = "kinesis.event.source"
+    )]
     kinesis_event_source: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", rename = "kinesis.event.source_arn")]
+    #[serde(
+        skip_serializing_if = "Option::is_none",
+        rename = "kinesis.event.source_arn"
+    )]
     kinesis_event_source_arn: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none", rename = "kafka.topic")]
     kafka_topic: Option<String>,
@@ -162,7 +168,7 @@ impl From<&process::MetadataContext> for JsonMessage {
             message: Value::Null,
             s3_key: mctx.get("s3.object.key"),
             s3_bucket: mctx.get("s3.bucket"),
-            cw_log_group:  mctx.get("cw.log.group"),
+            cw_log_group: mctx.get("cw.log.group"),
             cw_log_stream: mctx.get("cw.log.stream"),
             cw_owner: mctx.get("cw.owner"),
             kinesis_event_id: mctx.get("kinesis.event.id"),
@@ -174,7 +180,7 @@ impl From<&process::MetadataContext> for JsonMessage {
             ecr_scan_source: mctx.get("ecr.scan.source"),
             sqs_event_source: mctx.get("sqs.event.source"),
             sqs_event_id: mctx.get("sqs.event.id"),
-            
+
             // to be deprecated
             stream_name: mctx.get("cw.log.stream"),
             loggroup_name: mctx.get("cw.log.group"),
@@ -213,7 +219,11 @@ impl JsonMessage {
         }
     }
 
-    fn with_selected_metadata(mut self, mctx: &process::MetadataContext, selected_metadata_keys: Vec<&str>) -> Self {
+    fn with_selected_metadata(
+        mut self,
+        mctx: &process::MetadataContext,
+        selected_metadata_keys: Vec<&str>,
+    ) -> Self {
         for key in selected_metadata_keys {
             match key {
                 "s3.object.key" => self.s3_key = mctx.get("s3.object.key"),
@@ -222,9 +232,13 @@ impl JsonMessage {
                 "cw.log.stream" => self.cw_log_stream = mctx.get("cw.log.stream"),
                 "cw.owner" => self.cw_owner = mctx.get("cw.owner"),
                 "kinesis.event.id" => self.kinesis_event_id = mctx.get("kinesis.event.id"),
-                "kinesis.event.name" =>  self.kinesis_event_name = mctx.get("kinesis.event.name"),
-                "kinesis.event.source" => self.kinesis_event_source = mctx.get("kinesis.event.source"),
-                "kinesis.event.source_arn" => self.kinesis_event_source_arn = mctx.get("kinesis.event.source_arn"),
+                "kinesis.event.name" => self.kinesis_event_name = mctx.get("kinesis.event.name"),
+                "kinesis.event.source" => {
+                    self.kinesis_event_source = mctx.get("kinesis.event.source")
+                }
+                "kinesis.event.source_arn" => {
+                    self.kinesis_event_source_arn = mctx.get("kinesis.event.source_arn")
+                }
                 "kafka.topic" => self.kafka_topic = mctx.get("kafka.topic"),
                 "ecr.scan.id" => self.ecr_scan_id = mctx.get("ecr.scan.id"),
                 "ecr.scan.source" => self.ecr_scan_source = mctx.get("ecr.scan.source"),
@@ -348,7 +362,6 @@ fn convert_to_log_entry(
         Err(_) => Value::String(log),
     };
 
-
     let add_metadata: Vec<&str> = config.add_metadata.split(',').map(|s| s.trim()).collect();
     tracing::debug!("add_metadata: {:?}", add_metadata);
     let mut message = JsonMessage::new(msg).with_selected_metadata(mctx, add_metadata);
@@ -419,7 +432,6 @@ async fn send_logs(
     );
     Ok(())
 }
-
 
 // fn dynamic_metadata_value(mctx: process::MetadataContext, value: String) -> String {}