+++ src/metrics/mod.rs
@@ -1,13 +1,22 @@
-use crate::clients::AwsClients;
 use crate::events;
-use async_recursion::async_recursion;
+use crate::metrics::config::Config;
+use aws_lambda_events::firehose::KinesisFirehoseResponse;
 use lambda_runtime::{Error, LambdaEvent};
-use tracing::warn;
+use tracing::error;
 
-#[async_recursion]
+pub mod config;
+pub mod process;
+
 // metric telemetry handler
-// TODO: implement
-pub async fn handler(_: &AwsClients, _: LambdaEvent<events::Combined>) -> Result<(), Error> {
-    warn!("metrics telemetry mode not implemented");
-    Ok(())
+pub async fn handler(
+    config: &Config,
+    event: LambdaEvent<events::Combined>,
+) -> Result<KinesisFirehoseResponse, Error> {
+    match event.payload {
+        events::Combined::Firehose(firehose_event) => process::transform_firehose_event(config, firehose_event).await,
+        _ => {
+            error!("incompatible event type for metrics telemetry mode");
+            Err("incompatible event type for metrics telemetry mode".to_string().into())
+        }
+    }
 }