+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 ### 🧰 Bug fixes 🧰
 - Added support for dynamic allocation of Application and Subsystem names based on json key from log.
 
-### v1.2.0 / 2025-1-7
+### v1.3.0 / 2025-01-20
+### 💡 Enhancements 💡
+- New intergration workflow added for ingesting Cloudwatch Stream Metrics via Firehose over PrivateLink
+- Add Cloudwatch Metrics Stream creation to custom resource function
+
+### v1.2.0 / 2025-01-7
 ### 🧰 Bug fixes 🧰
 - Add permissions to custom lambda for `event-source-mapping`
 ### 💡 Enhancements 💡